refactor: test CLI args (#159)
Why?
It's too easy to make a config change and forget to make a corresponding
change to the CLI arg parsing logic.

What?
Move ArgumentParser from configs.helpers into configs.cli_args. Add a
test that checks that every config option has a corresponding CLI arg.
